requirements.txt
# Core Data Science Libraries
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.11.1

# Time Series and Forecasting
prophet==1.1.4
statsmodels==0.14.0
lightgbm==4.0.0
xgboost==1.7.6
catboost==1.2.2

# Deep Learning
torch==2.0.1
tensorflow==2.13.0
keras==2.13.1

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
bokeh==3.2.2

# Model Explainability
shap==0.42.1
lime==0.2.0.1
eli5==0.13.0

# Utilities
tqdm==4.65.0
joblib==1.3.2
pyyaml==6.0.1
python-dotenv==1.0.0

# Development and Testing
pytest==7.4.0
black==23.7.0
flake8==6.0.0
jupyter==1.0.0

# Data Processing
openpyxl==3.1.2
xlsxwriter==3.1.2

